Extracted DefaultSentenceBuilder to public class enabling derived types
